move models types and definitions into shared-types
