+++ apps/api/package.json
@@ -22,6 +22,7 @@
     "@elysiajs/openapi": "^1.4.11",
     "@hebo/database": "workspace:*",
     "@hebo/shared-api": "workspace:*",
+    "@hebo/shared-types": "workspace:*",
     "crypto-random-string": "^5.0.0",
     "elysia": "^1.4.9",
     "pg": "^8.16.3",
