+++ apps/gateway/package.json
@@ -25,6 +25,7 @@
     "@hebo/database": "workspace:*",
     "@hebo/shared-api": "workspace:*",
     "@hebo/shared-data": "workspace:*",
+    "@hebo/shared-types": "workspace:*",
     "ai": "^5.0.19",
     "elysia": "^1.4.9",
     "sst": "^3.17.13",
